requirements.txt
asgiref==3.11.0
Django==6.0
paho-mqtt==2.1.0
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.2
